+++ lib/models/myoro_table_column_width_configuration.dart
@@ -1,10 +1,10 @@
 import 'package:equatable/equatable.dart';
 import 'package:faker/faker.dart';
+import 'package:flutter/foundation.dart';
 import 'package:myoro_flutter_library/myoro_flutter_library.dart';
 
 /// Width configuration of a [MyoroTable].
-///
-/// TODO: Needs to be tested.
+@immutable
 class MyoroTableColumnWidthConfiguration extends Equatable {
   /// Width configuration enum.
   final MyoroTableColumnWidthConfigurationEnum typeEnum;