+++ lib/models/myoro_table_row.dart
@@ -1,4 +1,5 @@
 import 'package:equatable/equatable.dart';
+import 'package:faker/faker.dart';
 import 'package:flutter/widgets.dart';
 import 'package:myoro_flutter_library/myoro_flutter_library.dart';
 
@@ -8,8 +9,7 @@ import 'package:myoro_flutter_library/myoro_flutter_library.dart';
 typedef MyoroTableRowTapEvent<T> = void Function(T item);
 
 /// Non-title row of a [MyoroTable].
-///
-/// TODO: Needs to be tested.
+@immutable
 class MyoroTableRow<T> extends Equatable {
   /// Function executed when the [MyoroTableRow] is tapped.
   final MyoroTableRowTapEvent<T>? onTapDown;
@@ -24,6 +24,14 @@ class MyoroTableRow<T> extends Equatable {
   final List<Widget> cells;
 
   const MyoroTableRow({this.onTapDown, this.onTapUp, required this.cells});
+
+  MyoroTableRow.fake({int? cellQuantity})
+    : onTapDown = faker.randomGenerator.boolean() ? ((_) {}) : null,
+      onTapUp = faker.randomGenerator.boolean() ? ((_) {}) : null,
+      cells = List.generate(
+        cellQuantity ?? faker.randomGenerator.integer(10),
+        (int index) => Text('Cell #$index'),
+      );
 
   MyoroTableRow<T> copyWith({
     MyoroTableRowTapEvent<T>? onTapDown,
@@ -46,7 +54,7 @@ class MyoroTableRow<T> extends Equatable {
 
   @override
   String toString() =>
-      'MyoroTableRow(\n'
+      'MyoroTableRow<$T>(\n'
       '  onTapDown: $onTapDown,\n'
       '  onTapUp: $onTapUp,\n'
       '  cells: $cells,\n'