+++ lib/theme_extensions/tables/myoro_table_theme_extension.dart
@@ -5,8 +5,6 @@ import 'package:flutter/material.dart';
 import 'package:myoro_flutter_library/myoro_flutter_library.dart';
 
 /// [ThemeExtension] of [MyoroTable].
-///
-/// TODO: Needs to be tested.
 class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> {
   /// [Container.decoration] of [_MyoroTableState].
   final BoxDecoration decoration;
@@ -16,6 +14,9 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
 
   /// Spacing between columns.
   final double columnSpacing;
+
+  /// Default [TextStyle] of [_Row].
+  final TextStyle rowTextStyle;
 
   /// [EdgeInsets] of [_Loader], [_EmptyMessage] and [_ErrorMessage].
   final EdgeInsets loaderEmptyMessageErrorMessagePadding;
@@ -30,6 +31,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
     required this.decoration,
     required this.columnTextStyle,
     required this.columnSpacing,
+    required this.rowTextStyle,
     required this.loaderEmptyMessageErrorMessagePadding,
     required this.emptyMessageTextStyle,
     required this.errorMessageTextStyle,
@@ -49,6 +51,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
       ),
       columnTextStyle: typography.randomTextStyle,
       columnSpacing: faker.randomGenerator.decimal(scale: 20),
+      rowTextStyle: typography.randomTextStyle,
       loaderEmptyMessageErrorMessagePadding: EdgeInsets.all(
         faker.randomGenerator.decimal(scale: 50),
       ),
@@ -65,6 +68,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
       ),
       columnTextStyle = textTheme.titleMedium!,
       columnSpacing = 10,
+      rowTextStyle = textTheme.bodyMedium!,
       loaderEmptyMessageErrorMessagePadding = const EdgeInsets.all(10),
       emptyMessageTextStyle = textTheme.headlineMedium!,
       errorMessageTextStyle = textTheme.headlineMedium!.withColor(colorScheme.error);
@@ -74,6 +78,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
     BoxDecoration? decoration,
     TextStyle? columnTextStyle,
     double? columnSpacing,
+    TextStyle? rowTextStyle,
     EdgeInsets? loaderEmptyMessageErrorMessagePadding,
     TextStyle? emptyMessageTextStyle,
     TextStyle? errorMessageTextStyle,
@@ -82,6 +87,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
       decoration: decoration ?? this.decoration,
       columnTextStyle: columnTextStyle ?? this.columnTextStyle,
       columnSpacing: columnSpacing ?? this.columnSpacing,
+      rowTextStyle: rowTextStyle ?? this.rowTextStyle,
       loaderEmptyMessageErrorMessagePadding:
           loaderEmptyMessageErrorMessagePadding ?? this.loaderEmptyMessageErrorMessagePadding,
       emptyMessageTextStyle: emptyMessageTextStyle ?? this.emptyMessageTextStyle,
@@ -99,6 +105,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
       decoration: BoxDecoration.lerp(decoration, other.decoration, t),
       columnTextStyle: TextStyle.lerp(columnTextStyle, other.columnTextStyle, t),
       columnSpacing: lerpDouble(columnSpacing, other.columnSpacing, t),
+      rowTextStyle: TextStyle.lerp(rowTextStyle, other.rowTextStyle, t),
       loaderEmptyMessageErrorMessagePadding: EdgeInsets.lerp(
         loaderEmptyMessageErrorMessagePadding,
         other.loaderEmptyMessageErrorMessagePadding,
@@ -116,6 +123,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
         other.decoration == decoration &&
         other.columnTextStyle == columnTextStyle &&
         other.columnSpacing == columnSpacing &&
+        other.rowTextStyle == rowTextStyle &&
         other.loaderEmptyMessageErrorMessagePadding == loaderEmptyMessageErrorMessagePadding &&
         other.emptyMessageTextStyle == emptyMessageTextStyle &&
         other.errorMessageTextStyle == errorMessageTextStyle;
@@ -127,6 +135,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
       decoration,
       columnTextStyle,
       columnSpacing,
+      rowTextStyle,
       loaderEmptyMessageErrorMessagePadding,
       emptyMessageTextStyle,
       errorMessageTextStyle,
@@ -139,6 +148,7 @@ class MyoroTableThemeExtension extends ThemeExtension<MyoroTableThemeExtension> 
       '  decoration: $decoration,\n'
       '  columnTextStyle: $columnTextStyle,\n'
       '  columnSpacing: $columnSpacing,\n'
+      '  rowTextStyle: $rowTextStyle,\n'
       '  loaderEmptyMessageErrorMessagePadding: $loaderEmptyMessageErrorMessagePadding\n'
       '  emptyMessageTextStyle: emptyMessageTextStyle,\n'
       '  errorMessageTextStyle: $errorMessageTextStyle,\n'