+++ test/functions/theme_data/create_myoro_theme_extensions_test.dart
@@ -96,6 +96,9 @@ void main() {
       MyoroSnackBarContainerThemeExtension,
       MyoroSnackBarThemeExtension,
 
+      // Tables
+      MyoroTableThemeExtension,
+
       // Tooltips
       MyoroTooltipThemeExtension,
     ];